feat: request body 필드 및 파싱 로직 추가
