+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -1,5 +1,8 @@
 package org.apache.coyote.http11;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -12,13 +15,26 @@ public class HttpRequest {
     private String protocol;
     private final Map<String, String> headers = new HashMap<>();
     private final Map<String, String> queryParameters = new HashMap<>();
+    private String body;
 
-    public HttpRequest(List<String> request) {
-        String[] split = request.getFirst().split(" ");
+    public HttpRequest(BufferedReader reader) throws IOException {
+        List<String> input = getInput(reader);
+
+        String[] split = input.getFirst().split(" ");
         method = split[0];
         protocol = split[2];
         parseUri(split[1]);
-        parseHeaders(request.subList(1, request.size()));
+        parseHeaders(input.subList(1, input.size()));
+        parseBody(reader);
+    }
+
+    private List<String> getInput(BufferedReader reader) throws IOException {
+        List<String> lines = new ArrayList<>();
+        String line;
+        while ((line = reader.readLine()) != null && !line.isEmpty()) {
+            lines.add(line);
+        }
+        return lines;
     }
 
     private void parseUri(String uri) {
@@ -58,12 +74,27 @@ public class HttpRequest {
         );
     }
 
+    private void parseBody(BufferedReader reader) throws IOException {
+        String contentLengthValue = headers.get("Content-Length");
+        if (contentLengthValue == null) {
+            return;
+        }
+        int contentLength = Integer.parseInt(contentLengthValue);
+        char[] buffer = new char[contentLength];
+        reader.read(buffer, 0, contentLength);
+        this.body = new String(buffer);
+    }
+
     public String getMethod() {
         return method;
     }
 
     public String getPath() {
         return path;
+    }
+
+    public String getHeader(String key) {
+        return headers.get(key);
     }
 
     public String getQueryParameter(String key) {
